# backend/swagger.yaml
openapi: 3.0.3
info:
  title: Nuvana Production API
  description: |
    Complete API documentation for the Nuvana Education Platform backend.

    **Authentication**: All endpoints (except those marked Public) require a Bearer JWT token in the `Authorization` header.
    The `x-school-id` header is required for tenant isolation on most endpoints.

    **Roles**: super_admin, school_admin, teacher, student

    **Rate Limiting**: Global 100 req/min. AI endpoints: 5 req/min.
  version: 1.0.0
  contact:
    name: Nuvana Team
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.nuvana.in
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SchoolId:
      name: x-school-id
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant school ID extracted from JWT
  schemas:
...
    CreateGradeDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
    UpdateGradeDto:
      type: object
      properties:
        name:
          type: string
    CreateClassDto:
      type: object
      required: [name, grade_level_id]
      properties:
        name:
          type: string
        grade_level_id:
          type: integer
    UpdateClassDto:
      type: object
      properties:
        name:
          type: string
        grade_level_id:
          type: integer
    CreateSubjectDto:
      type: object
      required: [name]
      properties:
        name:
          type: string
...
    AssignTeacherToClassDto:
      type: object
      required: [teacher_id, class_id]
      properties:
        teacher_id:
          type: string
          format: uuid
        class_id:
          type: string
          format: uuid
...
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
security:
  - BearerAuth: []
paths:
...
  /academic/grades:
    post:
      tags: [Academic]
      summary: Create grade level (school_admin, super_admin)
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGradeDto'
      responses:
        '201':
          description: Grade created
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all grades (school_admin, teacher, super_admin)
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of grades
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/grades/{id}:
    patch:
      tags: [Academic]
      summary: Update grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGradeDto'
      responses:
        '200':
          description: Grade updated
          content:
            application/json:
              schema:
                type: object

    delete:
      tags: [Academic]
      summary: Delete grade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Grade deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - CLASSES ────────────────────────────────
  /academic/classes:
    post:
      tags: [Academic]
      summary: Create class
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassDto'
      responses:
        '201':
          description: Class created
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all classes
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of classes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/classes/{id}:
    patch:
      tags: [Academic]
      summary: Update class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClassDto'
      responses:
        '200':
          description: Class updated
          content:
            application/json:
              schema:
                type: object

    delete:
      tags: [Academic]
      summary: Delete class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Class deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
  /academic/classes/{classId}/students:
    get:
      tags: [Academic]
      summary: Get students by class (teacher, school_admin)
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of students in class
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

  # ── ACADEMIC - SUBJECTS ───────────────────────────────
  /academic/subjects:
    post:
      tags: [Academic]
      summary: Create master subject
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubjectDto'
      responses:
        '201':
          description: Subject created
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all master subjects
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/subjects/{id}:
    delete:
      tags: [Academic]
      summary: Delete subject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Subject deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - GRADE SUBJECTS ─────────────────────────
  /academic/grade-subjects:
    post:
      tags: [Academic]
      summary: Assign subjects to grade
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignSubjectsToGradeDto'
      responses:
        '201':
          description: Subjects assigned to grade
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all grade-subject mappings
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of grade-subject mappings
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/grade-subjects/grade/{gradeId}:
    get:
      tags: [Academic]
      summary: Get subjects by grade
      parameters:
        - name: gradeId
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Subjects for grade
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/grade-subjects/{id}:
    delete:
      tags: [Academic]
      summary: Delete grade-subject mapping
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - TEACHER-CLASS ASSIGNMENTS ──────────────
  /academic/teacher-classes:
    post:
      tags: [Academic]
      summary: Assign teacher to class
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignTeacherToClassDto'
      responses:
        '201':
          description: Teacher assigned
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all teacher-class assignments
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/teacher-classes/teacher/{teacherId}:
    get:
      tags: [Academic]
      summary: Get classes assigned to teacher (as class teacher)
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Teacher's classes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/teacher-classes/teacher/{teacherId}/all:
    get:
      tags: [Academic]
      summary: Get ALL classes where teacher teaches (class teacher + subject teacher)
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: All teaching classes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/teacher-classes/{id}:
    delete:
      tags: [Academic]
      summary: Delete teacher-class assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - TEACHER-SUBJECT ASSIGNMENTS ────────────
  /academic/teacher-subjects:
    post:
      tags: [Academic]
      summary: Assign subjects to teacher
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignSubjectsToTeacherDto'
      responses:
        '201':
          description: Subjects assigned
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all teacher-subject assignments
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/teacher-subjects/teacher/{teacherId}:
    get:
      tags: [Academic]
      summary: Get subjects by teacher
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Teacher's subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/teacher-subjects/teacher/{teacherId}/all:
    get:
      tags: [Academic]
      summary: Get ALL subjects by teacher (detailed with grade info)
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Teacher's detailed subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/teacher-subjects/{id}:
    delete:
      tags: [Academic]
      summary: Delete teacher-subject assignment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - FILE CATEGORIES ────────────────────────
  /academic/file-categories:
    post:
      tags: [Academic]
      summary: Create file category
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileCategoryDto'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all file categories
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/file-categories/{id}:
    patch:
      tags: [Academic]
      summary: Update file category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileCategoryDto'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                type: object

    delete:
      tags: [Academic]
      summary: Delete file category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Category deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - EXAM TYPES ────────────────────────────
  /academic/exam-types:
    post:
      tags: [Academic]
      summary: Create exam type
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExamTypeDto'
      responses:
        '201':
          description: Exam type created
          content:
            application/json:
              schema:
                type: object

    get:
      tags: [Academic]
      summary: Get all exam types
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: List of exam types
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /academic/exam-types/{id}:
    patch:
      tags: [Academic]
      summary: Update exam type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateExamTypeDto'
      responses:
        '200':
          description: Exam type updated
          content:
            application/json:
              schema:
                type: object

    delete:
      tags: [Academic]
      summary: Delete exam type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Exam type deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - TIMETABLE ──────────────────────────────
  /academic/timetable/class/{classId}:
    get:
      tags: [Academic]
      summary: Get weekly timetable for class
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Weekly timetable
          content:
            application/json:
              schema:
                type: object
  /academic/timetable/periods:
    post:
      tags: [Academic]
      summary: Create or update period
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePeriodDto'
      responses:
        '201':
          description: Period created/updated
          content:
            application/json:
              schema:
                type: object
  /academic/timetable/periods/{id}:
    patch:
      tags: [Academic]
      summary: Update period
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePeriodDto'
      responses:
        '200':
          description: Period updated
          content:
            application/json:
              schema:
                type: object

    delete:
      tags: [Academic]
      summary: Delete period
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Period deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'

  # ── ACADEMIC - HELPERS ────────────────────────────────
  /academic/helper/grade-subject/{classId}/{subjectName}:
    get:
      tags: [Academic]
      summary: Get grade_subject_id by class and subject name
      parameters:
        - name: classId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: subjectName
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Grade subject ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
  /academic/helper/exam-type/{examTypeName}:
    get:
      tags: [Academic]
      summary: Get exam_type_id by name
      parameters:
        - name: examTypeName
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/SchoolId'
      responses:
        '200':
          description: Exam type ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer

  # ── ACADEMIC - MANUAL MARKS ───────────────────────────
  /academic/manual-marks:
    post:
      tags: [Academic]
      summary: Save manual marks (teacher)
      parameters:
        - $ref: '#/components/parameters/SchoolId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveManualMarksDto'
      responses:
        '201':
          description: Marks saved
          content:
            application/json:
              schema:
                type: object

  # ══════════════════════════════════════════════════════
  #  ATTENDANCE
  # ══════════════════════════════════════════════════════
...